Extracted inner CaseSummary implementation from CaseFilteringService.
